#9 fix use local storage hook
add to distance radio button form
